third_party/WebKit/LayoutTests/fast/forms/radio/radio-group-remove-required.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<form>
<input type=radio name=requiredRadio id=rbutton required>
</form>
<script>
description('Removing a named required radio button shouldn\'t assert');

var radio = document.getElementById('rbutton');
shouldBeTrue('document.getElementsByName("requiredRadio")[0].required');
shouldBeTrue('radio.required');
radio.parentElement.removeChild(radio);
shouldBeTrue('radio.required');
</script>
</body>
</html>
